pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "expforge"
version = "0.1.0"
description = "Lightweight experiment scaffolding for ML workflows"
readme = "README.md"
authors = [{name = "Your Name"}]
requires-python = ">=3.13"
dependencies = [
    "typer[all]>=0.9",
    "pillow>=9.0",
    "numpy>=1.23",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "tensorflow>=2.11.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-storage>=2.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]
